+++ CHANGES.md
@@ -11,6 +11,8 @@
 
 ## develop
 
+- [ADD] `SampleTableAccessor::get_sample_by_timestamp()` を追加する
+   - @sile
 - [ADD] `SampleAccessor::timestamp()` を追加する
    - @sile
 - [ADD] `SampleAccessor::sync_sample_index()` を追加する